+++ src/frontend/app/layout.tsx
@@ -3,6 +3,7 @@ import { Inter } from "next/font/google";
 import "./globals.css";
 import Sidebar from "@/components/Sidebar";
 import Header from "@/components/Header";
+import Footer from "@/components/Footer";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -18,16 +19,30 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={`${inter.className} flex items-start justify-between`}>
+      <body className={`${inter.className} min-h-screen flex flex-col`}>
         
-        <div className="hidden md:flex md:shrink-10 min-w-[300px] border-r min-h-screen">
-          <Sidebar/>
+        <div className="flex-grow grid md:grid-cols-[300px_1fr]">
+
+          <div className="h-full hidden md:block md:shrink-10 min-w-[300px] border-r">
+            <Sidebar/>
+          </div>
+
+          <main className="ml-0">
+
+            <Header/>
+
+            <div className="grid w-full">
+              {children}
+            </div>
+            
+          </main>
+        
         </div>
 
-        <main className="grid w-full h-full">
-          <Header/>
-          {children}
-        </main>
+        <footer className="w-full">
+          <Footer/>
+        </footer>
+
       </body>
     </html>
   );