+++ src/frontend/components/ComponentOverviewDataTable.tsx
@@ -60,7 +60,7 @@ export function ComponentOverviewDataTable<TData, TValue>({
                 data-state={row.getIsSelected() && "selected"}
               >
                 {row.getVisibleCells().map((cell) => (
-                  <TableCell key={cell.id}>
+                  <TableCell key={cell.id} className="text-sm justify-items-start text-left">
                     {flexRender(cell.column.columnDef.cell, cell.getContext())}
                   </TableCell>
                 ))}
@@ -68,7 +68,7 @@ export function ComponentOverviewDataTable<TData, TValue>({
             ))
           ) : (
             <TableRow>
-              <TableCell colSpan={columns.length} className="h-24 text-center">
+              <TableCell colSpan={columns.length} className="h-8 justify-items-start text-left">
                 No results.
               </TableCell>
             </TableRow>