+++ src/frontend/components/Header.tsx
@@ -6,7 +6,7 @@ import AppMenu from "./AppMenu";
 export default function Header() {
   return (
     <>
-      <div className="flex gap-4 p-4 border-b font-bold">
+      <div className="flex gap-4 p-4 border-b font-bold grow">
         <BookOpenText /><h1>Catalogue of Second Chances</h1>
       </div>